pyproject.toml
[project]
name = "qbo-mcp"
version = "0.1.0"
description = "MCP server for QuickBooks Online integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "intuit-oauth>=1.2.6",
    "mcp[cli]>=1.6.0",
    "python-quickbooks>=0.9.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
qbo-mcp = "qbo_mcp:main"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
